authentication/ms-sso/requirements.txt
pip install fastapi-azure-auth

fastapi
uvicorn
fastapi-sso
python-dotenv
httpx
itsdangerous